refactor: 데이터 베이스 생성 부분 모듈로 이동 #46
