+++ app/src/main/java/com/kdjj/ratatouille/di/data/DataBaseModule.kt
@@ -1,12 +1,19 @@
 package com.kdjj.ratatouille.di.data
 
 import android.content.Context
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import androidx.sqlite.db.SupportSQLiteDatabase
 import com.kdjj.local.database.RecipeDatabase
+import com.kdjj.local.model.RecipeTypeEntity
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import javax.inject.Singleton
 
 @Module
@@ -15,8 +22,22 @@ object DataBaseModule {
 	
 	@Provides
 	@Singleton
-	fun provideRecipeDataBase(@ApplicationContext context: Context) =
-		RecipeDatabase.getInstance(context)
+	fun provideRecipeDataBase(@ApplicationContext context: Context): RecipeDatabase {
+		return Room.databaseBuilder(
+			context.applicationContext,
+			RecipeDatabase::class.java,
+			RecipeDatabase.RECIPE_DATABASE_NAME
+		).addCallback(object : RoomDatabase.Callback() {
+			override fun onCreate(db: SupportSQLiteDatabase) {
+				super.onCreate(db)
+				CoroutineScope(Dispatchers.IO).launch {
+					RecipeTypeEntity.defaultTypes.forEach { recipeType ->
+						db.execSQL("INSERT INTO RecipeType VALUES (${recipeType.recipeTypeId}, '${recipeType.title}');")
+					}
+				}
+			}
+		}).build()
+	}
 	
 	@Provides
 	@Singleton
